+++ src/main/java/com/testingSystem/service/AnswerService.java
@@ -4,5 +4,5 @@ import com.testingSystem.entity.Answer;
 import org.springframework.stereotype.Service;
 
 @Service
-public interface AnswerService extends EntityService<Answer, Long> {
+public interface AnswerService extends EntityService<Answer, Integer> {
 }
